cms-ckan-sync/requirements.txt
# Core dependencies
requests>=2.31.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
tenacity>=8.2.0

# Web UI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6
httpx>=0.25.0

# Scheduling
apscheduler>=3.10.0
pytz>=2023.3

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
